Free app folder from other inner folders...moved to src
