+++ src/pages/dummyHome/home.jsx
@@ -1,7 +1,7 @@
 /* eslint-disable arrow-body-style */
 import React from 'react';
 import { Link } from 'react-router-dom';
-import FooterContainer from '../../../components/footer/FooterContainer'
+import FooterContainer from '../../components/footer/FooterContainer'
 
 const Home = () => (
   <>
