+++ src/routes/index.jsx
@@ -2,14 +2,14 @@ import React from 'react'
 import { Routes, Route } from 'react-router-dom'
 import Home from '../pages/home/homeContainer'
 import About from '../pages/dummyAbout/about'
-import BackOfficeContainer from '../../components/BackOffice/BackOfficeContainer'
-import HomeFormContainer from '../../components/Forms/HomeForm/HomeFormContainer'
-import LoginFormContainer from '../../components/Forms/LoginForm/LoginFormContainer'
-import RegisterFormContainer from '../../components/Forms/RegisterForm/RegisterFormContainer'
-import NewsContainer from '../../components/News/NewsContainer'
-import NewsByIdContainer from '../../components/News/NewByIdContainer'
-import BackOficceUsers from '../../components/BackOffice/users/UsersContainer'
-import MyProfileContainer from '../../components/MyProfile/MyProfileContainer'
+import BackOfficeContainer from '../components/BackOffice/BackOfficeContainer'
+import HomeFormContainer from '../components/Forms/HomeForm/HomeFormContainer'
+import LoginFormContainer from '../components/Forms/LoginForm/LoginFormContainer'
+import RegisterFormContainer from '../components/Forms/RegisterForm/RegisterFormContainer'
+import NewsContainer from '../components/News/NewsContainer'
+import NewsByIdContainer from '../components/News/NewByIdContainer'
+import BackOficceUsers from '../components/BackOffice/users/UsersContainer'
+import MyProfileContainer from '../components/MyProfile/MyProfileContainer'
 
 const Router = () => (
   <Routes>
